+++ src/generator.ts
@@ -1,5 +1,6 @@
 import * as fs from 'fs';
 import {gitLogOneLine} from "./gitUtils";
+import {getCommitMessagesObjByType} from "./commitLint";
 
 const generateChangelog = async (changelogFile: string, newVersion: string, newVersionTitle: string, fromTag: string = 'latest') => {
 
@@ -18,11 +19,15 @@ const generateChangelog = async (changelogFile: string, newVersion: string, newV
 
     try {
         const logLines: string[] = await gitLogOneLine(fromTag);
+        const logTypeMap = getCommitMessagesObjByType(logLines);
+        const sections: string[] = [];
+        Object.keys(logTypeMap).sort().forEach((type) => {
+            sections.push(`### ${type} \n\n- ${logTypeMap[type].join('\n- ')}`);
+        });
+        changelog = changelog.replace(
+            '# Changelog \n\n', `# Changelog \n\n## ${newVersion} ${newVersionTitle} \n${sections.join('\n\n')}\n`
+        );
 
-        console.log('>>', changelog);
-        changelog = changelog.replace('# Changelog \n\n', `# Changelog \n\n## ${newVersion} ${newVersionTitle} \n${logLines.join('\n')}\n`);
-
-        console.log(changelog);
         fs.writeFileSync(changelogFile, changelog, { flag: 'w+', mode: 0o666, flush: true });
     } catch (err) {
         console.error(err);