+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog 
 
 ## 1.0.0 2021-01-01 
-219bc53 feat(commitlint): add commitlint dep and config
-cf9c505 feat(all): initial commit
+### Code Refactoring 
+
+- 92f036f refactor(git): rename utils to git utils
+
+### Features 
+
+- 219bc53 feat(commitlint): add commitlint dep and config
+- cf9c505 feat(all): initial commit
+
+### Others 
+
+- b224f0f chore: update demo